requirements.txt
streamlit
sentence-transformers
faiss-cpu
python-dotenv
langchain-groq
numpy
torch
huggingface-hub